Integers Only
int - city tourism modifier

int tourism from city culture

war score ratio, small gifts, vote ratio, happiness factor

int - ai gold gifts

int - safe decimal

int - tile ratios

int - economic ratio

int - spy

int - economic and espionage

remove more floats

int - remove useless terrain floats

projects

int - military

int - gold gifts

int - quests

int - power ratio

int - capped gold gift

int - huts

int - city damage

int - diplomatic influence

int - player stuff

int - cities

int - policies, random, pregame

int - missions???

int - tech

int - Mayan Calendar

int - unit maintenance/ supply

decimal

combat

int - random stuff

rename

int - names

int - units

int - damage ratio

int - units

int - nukes

delay

int - citizens, votes

remove 0.f

int - trigger words, aka "double" in comments

double - border expansion

city strategy

int - unnecessary doubles tourism

fix tourism

int - city strategy

int - internet defense

int - danger plots

AI

int - square root

int- ai map flavor

int - dot product

int - round

int - flavor

int - plot vision

int - cleanup displacements

int - visibility optimizations

int - stopwatch

int - flavor

int - trade, iPow

int - unit damage

remove AUI_VOTING_SCORE_VOTING_CHOICE_PLAYER_ADJUST_FOR_FPTP

int - 0.

int - damage

int - constants

int - more constants

int - false alarms

fix num policy displayed

int - policy cost

fix purchase costs

remove trigger words

restore some square roots

Create behavior.txt

fix local vs global

fix sqrt for low numbers

clean up bad code

fix serious bug

fix w1 warnings

+++ DLC/LEKMOD_V28.2/nqmod-vs2008-master/CvGameCoreDLL_Expansion2/CvDiplomacyAI.h
@@ -52,7 +52,7 @@ struct MinorGoldGiftInfo
 {
 	PlayerTypes eMinor;
 	PlayerTypes eMajorRival;
-	float numSmallGiftsNeeded;
+	int numSmallGiftsNeeded;
 };
 class CvDiplomacyAI
 {
